moved Information folder to shared: ignored ESLint warnings related to @mui/icons and hex color
