+++ ui/components/MesheryPatterns/MesheryPatterns.tsx
@@ -58,7 +58,7 @@ import { modifyRJSFSchema } from '../../utils/utils';
 import { Edit as EditIcon } from '@mui/icons-material';
 import { updateVisibleColumns } from '../../utils/responsive-column';
 import { useWindowDimensions } from '../../utils/dimension';
-import InfoModal from '../General/Modals/Information/InfoModal';
+import InfoModal from '../shared/Modal/Information/InfoModal';
 import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import { DefaultTableCell, SortableTableCell } from '../connections/common';
 import DefaultError from '../General/error-404/index';
