+++ ui/components/SpacesSwitcher/DesignViewListItem.tsx
@@ -14,7 +14,7 @@ import {
   Typography,
   VisibilityChipMenu,
 } from '@sistent/sistent';
-import { VIEW_VISIBILITY } from '../General/Modals/Information/InfoModal';
+import { VIEW_VISIBILITY } from '../shared/Modal/Information/InfoModal';
 import {
   StyledAvatarContainer,
   StyledListIcon,
